requirements.txt
paho-mqtt==1.5.1
pymongo==3.11.4
click==8.0.1
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.19.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
psycopg2-binary==2.9.1